src/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
pymysql
chromadb
aiofiles
google-genai==1.72.0
argon2-cffi
python-jose[cryptography]
cryptography
pymupdf
types-aiofiles
types-python-jose
types-PyMySQL